Extended clean-up after to include oracle-aa artifacts

+++ passes/scripts/run_me.sh
@@ -7,12 +7,19 @@ if test "$installDir" == "" ; then
 fi
 
 # Install
-rm -rf build/ ; 
-mkdir build ; 
+# rm -rf build/ ; 
+
+if [ ! -d "build" ]
+then 
+  mkdir build ; 
+fi
 cd build ; 
-cmake3 -DCMAKE_INSTALL_PREFIX="${installDir}" -DCMAKE_BUILD_TYPE=Debug ../ ; 
+cmake -DCMAKE_INSTALL_PREFIX="${installDir}" -DCMAKE_BUILD_TYPE=Debug ../ ; 
 make ;
 make install ;
 cd ../
 
-ln -s build/compile_commands.json
+if [ ! -f "compile_commands.json" ]
+then 
+  ln -s build/compile_commands.json
+fi
